backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for baby stats"
authors = [
    { name = "Quentin Vanderlinden", email = "quentin.vanderlinden@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.1",
    "psycopg>=3.2.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.1",
    "ruff>=0.12.9",
    "django-stubs[compatible-mypy]>=5.2.0",
]

[tool.mypy]
disallow_untyped_defs = true
plugins = ["pydantic.mypy", "mypy_django_plugin.main"]
exclude = ["src/mysite/settings.py", "manage.py"]

[tool.django-stubs]
django_settings_module = "mysite.settings"

[tool.ruff]
exclude = []
